add CommentWrapper component
